% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HBSmDataImp.R
\name{HBSmDataImp}
\alias{HBSmDataImp}
\title{Hybrid Blau Model Data Import}
\usage{
HBSmDataImp(dataDir, nDim = 3, entities, entlist, d1lab, d2lab, d3lab = NULL)
}
\arguments{
\item{dataDir}{Data object you would like the function to use. Needs to be a data frame}

\item{nDim}{Dimensions of your matrix. Currently the model is capable of handling two or three dimensional matrices (default = 3)}

\item{entities}{Number of organizations in the model (default = 16)}

\item{entlist}{list of the organization names/types}

\item{d1lab}{Label for dimension 1}

\item{d2lab}{Label for dimension 2}

\item{d3lab}{Label for dimension 3}
}
\value{
Object list of data and necessary objects for the simulation
}
\description{
This function runs the data import script that allows users to import data that is readable by the simulation
}
\examples{
datalist <- HBSmDataImp(dataDir = HBSmexampledat,
nDim = 3,
entities=16,
entlist=orgnames,
d1lab = "educ",
d2lab = "age",
d3lab = "ic_income")
}
