% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HBSmSim.R
\name{HBSmSim}
\alias{HBSmSim}
\title{Hybrid Blau Model Sim}
\usage{
HBSmSim(
  glob = 10,
  chec = 500,
  maxr = 4000,
  tol = 8,
  nrange = 1,
  convr = 15,
  imp = 1,
  bin = 1,
  HBSmdata,
  prefix = "",
  year = "",
  saveDir,
  notif = TRUE
)
}
\arguments{
\item{glob}{Total number of runs that the model simulation will complete (default = 10)}

\item{chec}{Number of iterations before convergence check (default = 500). If no convergence is reached, the model will continue until it reaches the values set for maxr or until it finally does converge}

\item{maxr}{Number of runs until model is forced to stopped if no convergence is reached (default = 4000)}

\item{tol}{How little difference between org arrays there can be before the model is considered converged (default = 8)}

\item{nrange}{Range of the neighborhood cells. A neighborhood cell is considered to be the cells surrounding the focal cell that are within range if those value (default = 1). If r = 1 then there is one focal cell and 8 neighborhood cells.}

\item{convr}{Number of permutations for convergence check}

\item{imp}{Toggle for import data}

\item{bin}{Toggle for if any bin variables are on}

\item{HBSmdata}{Set this to equal the object you generated with the data import function}

\item{prefix}{prefix to your saveout data}

\item{year}{year of your saveout data}

\item{saveDir}{directory for where your data is saved to}

\item{notif}{Plays a sound when model is done. Set to TRUE or FALSE.}
}
\value{
The results of the model. The list outputlist has the relevant outputs of the model while objectdump contains objects used to run the model.
}
\description{
This function runs the main simulation model while allowing you to edit the simulations global parameters.
}
\examples{
HBSmSim(glob = 1,
chec = 200,
maxr = 500,
tol = 8,
nrange = 1,
convr = 15,
HBSmdata = datalist,
prefix="GSS",
year = "1974",
saveDir = saveDir)
}
