% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HBSmDataImp.R
\name{HBSmDataImp}
\alias{HBSmDataImp}
\title{Hybrid Blau Model Data Import}
\usage{
HBSmDataImp(dataDir, nDim = 3, entities, entlist, d1lab, d2lab, d3lab)
}
\arguments{
\item{dataDir}{Location/Directory of the data you are loading in}

\item{nDim}{Dimensions of your matrix. Currently the model is capable of handling two or three dimensional matrices (default = 3)}

\item{entities}{Number of organizations in the model (default = 16)}

\item{entlist}{list of the organization names/types}

\item{d1lab}{Label for dimension 1}

\item{d2lab}{Label for dimension 2}

\item{d3lab}{Label for dimension 3}
}
\value{
Imported data
}
\description{
This function runs the data import script that allows users to import data that is readable by the simulation
}
\examples{
urndataimp(dataDir = "C:/Users/username/Desktop/UrnModel/Data/GSS Cleaned", nDim = 3, entities=16, entlist=orgnames, d1lab = "educ", d2lab = "age", d3lab = "ic_income")
}
